+++ morpha/morpha/models/__init__.py
@@ -7,11 +7,17 @@ from .move_record import MoveRecord
 from .player_record import PlayerRecord
 from .pokemon_record import PokemonRecord
 from .battle_log import BattleLog
+from .pokemon import Pokemon
+from .player import Player
+from .battle import Battle
 
-__all__ = ['BattleLog',
+__all__ = ['Battle',
+           'BattleLog',
            'DamageRecord',
            'IRecord',
            'MoveRecord',
+           'Player',
            'PlayerRecord',
+           'Pokemon',
            'PokemonRecord',
            'utilities']