+++ morpha/documentation/release_notes.md
@@ -4,4 +4,6 @@
 ##### Features
 - Added logging.
 - Added BattleMetricsService (`ME-630`).
-- Added BattleLog, IRecord, and concrete Record models (`ME-630`).
+- Added Battle, Player, and Pokemon models (`ME-630`).
+- Added BattleLog model (`ME-630`).
+- Added IRecord and concrete Record models (`ME-630`).